// isms-core-platform/frontend-angular/src/app/shared/types.ts
export interface ProjectRead {
  id: string
  name: string
  description: string | null
  product_family: string
  status: string
  owner: string | null
  created_at: string
  updated_at: string
  organisation_name: string | null
}